Fixed readEventFile() call in getNextEvent(). Tested and compared same event read from xml and binary, and they match. Full simulation run not tested yet.

+++ software/alpide_dataflow_sim/src/event/EventBinary.cpp
@@ -196,7 +196,7 @@ void EventBinary::readChip(std::string event_filename,
   uint16_t col, row;
   bool done = false;
 
-  event_file.read((char*)&chip_id, sizeof(uint16_t));
+  event_file.read((char*)&chip_id, sizeof(uint8_t));
 
   ITS::detectorPosition pos = {layer_id, stave_id, mod_id, chip_id};
   unsigned int global_chip_id = ITS::detector_position_to_chip_id(pos);
